+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [v0.42.2](https://github.com/k1LoW/octocov/compare/v0.42.1...v0.42.2) - 2022-10-08
 - Use tagpr by @k1LoW in https://github.com/k1LoW/octocov/pull/173
+- Support for inserting code metrics report to body of pull request by @k1LoW in https://github.com/k1LoW/octocov/pull/172
 
 ## [v0.42.1](https://github.com/k1LoW/octocov/compare/v0.42.0...v0.42.1) (2022-09-28)